requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.0
SQLAlchemy==2.0.27
loguru==0.7.2 